+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on "Keep a Changelog".  This project adheres to Semantic Versioning.
 
 
+## [0.1.2] - 2019-06-12
+### Changed
+- GitClass.__init__:  Added attribute for git.Repo.init class.
+- GitMerge.priority_merge:  Captured standard error out in the exception handler.
+
+### Added
+- GitConfig:  Create a create a git.Repo.init class.
+- GitClass.create_init:  Create a git.Repo.init instance.
+
+
 ## [0.1.1] - 2019-06-04
 ### Added
 - __init__.py:  Added file for class use.